feat: move walletConnect, create new screen
