+++ src/handlers/deeplinks.ts
@@ -28,7 +28,7 @@ import { logger } from '@/logger';
 import {
   pair as pairWalletConnect,
   setHasPendingDeeplinkPendingRedirect,
-} from '@/utils/walletConnect';
+} from '@/walletConnect';
 import { analyticsV2 } from '@/analytics';
 import { FiatProviderName } from '@/entities/f2c';
