+++ src/screens/TransactionConfirmationScreen.js
@@ -111,7 +111,7 @@ import {
   SIGN_TYPED_DATA,
   SIGN_TYPED_DATA_V4,
 } from '@/utils/signingMethods';
-import { handleSessionRequestResponse } from '@/utils/walletConnect';
+import { handleSessionRequestResponse } from '@/walletConnect';
 import { isAddress } from '@ethersproject/address';
 import { logger, RainbowError } from '@/logger';
